Replace Settings with Organization Details
