+++ src/static-apps.js
@@ -17,8 +17,8 @@ export const staticApps = new Map(
   Object.entries({
     apps: staticApp('apps', 'App Center'),
     home: staticApp('home', 'Home', '/'),
+    organization: staticApp('organization', 'Organization'),
     permissions: staticApp('permissions', 'Permissions'),
-    settings: staticApp('settings', 'Settings'),
   })
 )
